fix: switch SetPlanningParams to all-float64 fields
foxglove_bridge accepts the standalone-float64 services we already use
(PreviewPlan, GetOutlinePath, etc.) but rejects mixed int32+float64
service requests at the rmw layer with "rmw_serialize: invalid data
size, typesupport identifier rosidl_typesupport_cpp not supported".
The actual ROS service is fine — direct ros2 service call works — but
the bridge can't relay the request from the GUI.

Sidestep the issue entirely: declare every wire field as float64 and
have the C++ handler cast outline_passes back to int. Matches the
shape of the other working services in mowgli_interfaces.

Refs #54

+++ ros2/src/mowgli_map/src/map_server_node.cpp
@@ -3434,9 +3434,11 @@ void MapServerNode::on_set_planning_params(
   // cached planner parameter.
   std::vector<std::string> changed;
 
-  if (req->outline_passes >= 0)
+  // outline_passes is wire-typed as float64 (see srv comment) so we cast
+  // back to int here. Negative values are the "unchanged" sentinel.
+  if (req->outline_passes >= 0.0)
   {
-    outline_passes_ = req->outline_passes;
+    outline_passes_ = static_cast<int>(req->outline_passes);
     changed.push_back("outline_passes=" + std::to_string(outline_passes_));
   }
   if (req->outline_offset >= 0.0)
